fix: Preserve existing branch status checks
In my previous PR for this
(https://github.com/cerebruminc/github-action-repo-settings-sync/pull/7)
I removed the checks value, however this results in an error when
querying the GitHub API ("checks" is required). My updated approach is to first retrieve the
existing set of checks on the repo, then when I update the branch
protection rules, keep the checks the same. I've tested this with the
cortex-api repo and it works as expected.

Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>